move c to src for appveyor
