+++ src/k.h
